requirements.txt
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-http>=1.27.0
pydantic-settings>=2.3.0
